MET-124: Split models
