+++ src/Akeneo/Tool/Bundle/MeasureBundle/Resources/public/model/direction.ts
@@ -1,0 +1,6 @@
+enum Direction {
+  Ascending = 'Ascending',
+  Descending = 'Descending',
+}
+
+export {Direction};
