+++ src/Akeneo/Tool/Bundle/MeasureBundle/Resources/public/model/label-collection.ts
@@ -1,0 +1,5 @@
+type LabelCollection = {
+  [locale: string]: string;
+};
+
+export {LabelCollection};
