+++ src/Akeneo/Tool/Bundle/MeasureBundle/Resources/public/model/operation.ts
@@ -1,0 +1,20 @@
+const MAX_OPERATION_COUNT = 5;
+
+enum Operator {
+  MUL = 'mul',
+  DIV = 'div',
+  ADD = 'add',
+  SUB = 'sub',
+}
+
+type Operation = {
+  operator: string;
+  value: string;
+};
+
+const emptyOperation = (): Operation => ({
+  value: '1',
+  operator: Operator.MUL,
+});
+
+export {MAX_OPERATION_COUNT, Operation, Operator, emptyOperation};
