+++ src/Akeneo/Tool/Bundle/MeasureBundle/Resources/public/shared/components/TextField.tsx
@@ -1,9 +1,10 @@
-import React, {ChangeEventHandler, useContext, forwardRef} from 'react';
+import React, {ChangeEventHandler, useContext} from 'react';
 import styled, {css} from 'styled-components';
 import {ValidationError} from 'akeneomeasure/model/validation-error';
 import {InputErrors} from 'akeneomeasure/shared/components/InputErrors';
 import {TranslateContext} from 'akeneomeasure/context/translate-context';
 import {Flag} from 'akeneomeasure/shared/components/Flag';
+import {useFocus} from 'akeneomeasure/shared/hooks/use-focus';
 
 const Input = styled.input.attrs(() => ({className: 'AknTextField'}))<{invalid: boolean}>`
   border: 1px solid ${props => props.theme.color.grey80};
@@ -23,6 +24,7 @@ type TextFieldProps = {
   locale?: string;
   required?: boolean;
   flag?: string;
+  autofocus?: boolean;
 
   value: string;
   onChange: ChangeEventHandler<Element>;
@@ -30,25 +32,40 @@ type TextFieldProps = {
   errors?: ValidationError[];
 };
 
-const TextField = forwardRef<HTMLInputElement, TextFieldProps & any>(
-  ({id, label, errors, propertyPath, required = false, flag, ...props}, ref) => {
-    const __ = useContext(TranslateContext);
+const TextField = ({
+  id,
+  label,
+  errors,
+  propertyPath,
+  required = false,
+  flag,
+  autofocus = false,
+  ...props
+}: TextFieldProps & any) => {
+  const __ = useContext(TranslateContext);
+  const [focusRef] = useFocus();
 
-    return (
-      <div className="AknFieldContainer">
-        <div className="AknFieldContainer-header">
-          <label className="AknFieldContainer-label" htmlFor={id}>
-            {label} {required && __('pim_common.required_label')}
-          </label>
-          {flag && <Flag localeCode={flag} />}
-        </div>
-        <div className="AknFieldContainer-inputContainer">
-          <Input ref={ref} id={id} type="text" autoComplete="off" invalid={errors && errors.length > 0} {...props} />
-        </div>
-        {errors && <InputErrors errors={errors} />}
+  return (
+    <div className="AknFieldContainer">
+      <div className="AknFieldContainer-header">
+        <label className="AknFieldContainer-label" htmlFor={id}>
+          {label} {required && __('pim_common.required_label')}
+        </label>
+        {flag && <Flag localeCode={flag} />}
       </div>
-    );
-  }
-);
+      <div className="AknFieldContainer-inputContainer">
+        <Input
+          ref={autofocus ? focusRef : undefined}
+          id={id}
+          type="text"
+          autoComplete="off"
+          invalid={errors && errors.length > 0}
+          {...props}
+        />
+      </div>
+      {errors && <InputErrors errors={errors} />}
+    </div>
+  );
+};
 
 export {TextField};
